requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pymongo==4.11.3
pydantic==2.11.3
python-dotenv==1.0.1
